Always initialize registry credentials
This should enable the use of cloud metadata services and docker configs on
local disk, without having to unnecessarily specify image pull secrets.

It was the case that k8schain was also pulling pull secrets from the
'default' service account in the kyverno namespace. I've switched this
to the 'kyverno' service account, which I think makes more sense.

Signed-off-by: Rob Best <robertbest89@gmail.com>